feat(domain): 구입 금액에 맞는 개수의 로또 생성 기능 추가

+++ src/main/java/lotto/domain/Lottos.java
@@ -11,7 +11,7 @@ public class Lottos {
         this.lottos = List.copyOf(generatedLottos);
     }
 
-    public static Lottos generateLottos(int count) {
+    public static Lottos generateLottos(long count) {
         List<Lotto> generatedLottos = new ArrayList<>();
         while (count-- > 0) {
             generatedLottos.add(new Lotto(LottoNumberGenerator.generate()));
